removed remaining references of snapshot-plugins repo, removed submodule, moved file used only by safesnap to its directoy
