+++ src/plugins/commentBox/plugin.json
@@ -2,6 +2,6 @@
   "name": "Comment Box",
   "version": "0.0.1",
   "author": "spiritbro1",
-  "website": "https://github.com/snapshot-labs/snapshot-plugins/tree/master/src/plugins/commentBox",
+  "website": "https://github.com/snapshot-labs/snapshot/tree/develop/src/plugins/commentBox",
   "icon": "ipfs://QmWpLpFpeQ3iH69uCeXXAfLu56YaL9mNESMGfeJ5XSpkhY"
 }
