+++ src/plugins/safeSnap/components/Input/MethodParameter.vue
@@ -1,5 +1,5 @@
 <script>
-import { isParameterValue } from '@/helpers/validator';
+import { isParameterValue } from '../../utils/validator';
 import { isArrayParameter } from '../../index';
 import SafeSnapInputAddress from './Address.vue';
 import SafeSnapInputArrayType from './ArrayType.vue';
